requirements.txt
asgiref==3.4.1
Django==4.0
gunicorn==20.1.0
sqlparse==0.4.2
tzdata==2021.5
psycopg2-binary==2.9.3


atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.4

iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.5.2
toml==0.10.2

factory-boy==3.2.1